move legacy/EOF traits and types to inspector::utils
